fixed bug with adding suites and part of a bug with deleting

+++ MFCTRSui/TestsTimerDialog.cpp
@@ -70,12 +70,13 @@ void TestsTimerDialog::OnAddClicked()
 {
 	if (clock_dlg.DoModal() == IDCANCEL)
 		return;
+	vector<SuiteRoot> selected_suites = clock_dlg.get_selected_suites();
 
-	ClockInstance item{ clock_dlg.get_clock_collection().front().get_suites(), clock_dlg.get_days(), clock_dlg.get_clock_name(), clock_dlg.get_tag(),
+	ClockInstance item{ clock_dlg.get_selected_suites(), clock_dlg.get_days(), clock_dlg.get_clock_name(), clock_dlg.get_tag(),
 		clock_dlg.get_name(), clock_dlg.get_threads(), clock_dlg.get_hour(), clock_dlg.get_minute(), clock_dlg.is_weekly(), UniqueNumber() };
 
 	list_items.push_back(item);
-	timersCollection.Add(item);
+//	timersCollection.Add(item);
 
 	AddToList(item);
 }
@@ -160,21 +161,29 @@ void TestsTimerDialog::OnEditClicked()
 	edit_clock_dlg.Init(list_items[selection].suites, vector<bool>(list_items[selection].suites.size(), true), 0, 0, 0);
 	if (edit_clock_dlg.DoModal() == IDOK)
 	{
-		timersCollection.Remove(list_items[selection]);
+//		timersCollection.Remove(list_items[selection]);
 
 		ClockInstance item{ edit_clock_dlg.get_clock_collection().front().get_suites(), edit_clock_dlg.get_days(), edit_clock_dlg.get_clock_name(), edit_clock_dlg.get_tag(),
 			edit_clock_dlg.get_name(), edit_clock_dlg.get_threads(), edit_clock_dlg.get_hour(), edit_clock_dlg.get_minute(), edit_clock_dlg.is_weekly(), UniqueNumber() };
 
 		ChangeListItem(item.clock_name, item.hour, item.minute, item.repeat, item.days, selection);
 
-		timersCollection.Add(item);
+//		timersCollection.Add(item);
 	}
 }
 
 
 void TestsTimerDialog::OnRemoveClicked()
 {
-	list_items.erase(list_items.begin() + selection);
+	if (selection >= list_items.size() || selection < 0)
+	{
+		logger << "selection is not within list_items range in TestsTimerDialog::OnRemoveClicked()";
+		return;
+	}
+
+	int asd = selection;
+	list_items.erase(list_items.begin() + asd);
+	// after DeleteItem system calls TestsTimerDialog::UpdateControls and assigns selection to -1
 	m_ListCtrl.DeleteItem(selection);
 }
